Uploading code coverage reports to coveralls.

+++ check-coverage.sh
@@ -23,6 +23,12 @@ FUNCTIONS_RATE=$(lcov --rc lcov_branch_coverage=1 --summary coverage2.info 2>&1|
 BRANCH_RATE=$(lcov --rc lcov_branch_coverage=1 --summary coverage2.info 2>&1|grep '^  branches'|sed 's/.*: //;s/%.*//')
 
 genhtml --branch-coverage -o build/coverage coverage2.info || exit 1
+
+if test "${COVERALLS_TOKEN}" != ""
+then
+    coveralls-lcov --repo-token "${COVERALLS_TOKEN}" coverage2.info || exit 1
+fi
+
 mv coverage2.info build/coverage.info
 rm coverage.info
